+++ CHANGELOG.md
@@ -52,9 +52,12 @@
 - Fixed digital state - issue #159
 - Fixed summary data - issue #160
 - Fixed an error in recorded max number of points - issue #162
+- Fix issue with summary when migrating from previous versions - issue $160
 
 - Updated the query editor layout
 - Added boundary type support in recorded values
 - Recognize partial usage of variables in elements
 - Added configuration to hide API errors in panel
 - Truncate time from grafana date time picker to seconds
+- Fixed warnings during deploy
+- Fixed LICENSE file